+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/domain/Command.kt
@@ -73,6 +73,7 @@ enum class Command {
     FDUNI,
     INVUNI,
 
+    ALIASVS,
     FDVS,
     INVVS,
 }